pre-map and pre-filter logs on the storage side
