+++ src/compute/src/logging/timely.rs
@@ -25,7 +25,6 @@ use mz_timely_util::replay::MzReplay;
 use timely::dataflow::Scope;
 use timely::dataflow::channels::pact::{ExchangeCore, Pipeline};
 use timely::dataflow::operators::Operator;
-use timely::dataflow::operators::core::Filter;
 use timely::dataflow::operators::generic::OutputBuilder;
 use timely::dataflow::operators::generic::builder_rc::OperatorBuilder;
 use timely::dataflow::operators::generic::operator::empty;
@@ -78,8 +77,7 @@ pub(super) fn construct(
             (empty(scope), token)
         };
 
-        // If we have a storage log reader, replay it, remap its IDs to avoid
-        // collisions with compute IDs, and concatenate with compute logs.
+        // If we have a storage log reader, replay it and concatenate with compute logs.
         let (logs, storage_token) = if let Some(reader) = storage_log_reader {
             use mz_timely_util::activator::RcActivator;
             use timely::dataflow::operators::Concatenate;
@@ -87,24 +85,6 @@ pub(super) fn construct(
             let activator = RcActivator::new("storage_timely_activator".to_string(), 128);
             let (storage_logs, s_token) =
                 [reader].mz_replay(scope, "storage timely logs", config.interval, activator);
-
-            let storage_logs = storage_logs
-                // Ignore storage park/unpark events.
-                .filter(|(_, x)| !matches!(x, TimelyEvent::Park { .. }))
-                // Remap storage IDs so they don't collide with compute IDs.
-                .unary(Pipeline, "Remap Storage IDs", |_cap, _info| {
-                    move |input, output| {
-                        input.for_each(|time, data| {
-                            output.session(&time).give_iterator(data.drain(..).map(
-                                |(t, mut event)| {
-                                    remap_timely_event_ids(&mut event);
-                                    (t, event)
-                                },
-                            ));
-                        });
-                    }
-                });
-
             let merged = scope.concatenate([logs, storage_logs]);
             (merged, Some(s_token))
         } else {
@@ -825,58 +805,5 @@ where
 {
     if vec.len() <= index {
         vec.resize(index + 1, Default::default());
-    }
-}
-
-/// Offset added to storage operator/channel IDs to avoid collisions with compute IDs.
-///
-/// This is large enough that compute IDs (which start from 0 and grow) will never reach it,
-/// but small enough to be representable as a `u64` with room for many storage operators.
-const STORAGE_ID_OFFSET: usize = 1 << 48;
-
-/// Remaps operator, channel, and address IDs in a `TimelyEvent` originating from a storage
-/// timely instance so they don't collide with compute's IDs.
-fn remap_timely_event_ids(event: &mut TimelyEvent) {
-    match event {
-        TimelyEvent::Operates(OperatesEvent { id, addr, .. }) => {
-            *id = id.wrapping_add(STORAGE_ID_OFFSET);
-            if let Some(first) = addr.first_mut() {
-                *first = first.wrapping_add(STORAGE_ID_OFFSET);
-            }
-        }
-        TimelyEvent::Channels(ChannelsEvent {
-            id,
-            scope_addr,
-            source,
-            target,
-            ..
-        }) => {
-            *id = id.wrapping_add(STORAGE_ID_OFFSET);
-            if let Some(first) = scope_addr.first_mut() {
-                *first = first.wrapping_add(STORAGE_ID_OFFSET);
-            }
-            source.0 = source.0.wrapping_add(STORAGE_ID_OFFSET);
-            target.0 = target.0.wrapping_add(STORAGE_ID_OFFSET);
-        }
-        TimelyEvent::Shutdown(ShutdownEvent { id }) => {
-            *id = id.wrapping_add(STORAGE_ID_OFFSET);
-        }
-        TimelyEvent::Schedule(ScheduleEvent { id, .. }) => {
-            *id = id.wrapping_add(STORAGE_ID_OFFSET);
-        }
-        TimelyEvent::Messages(MessagesEvent { channel, .. }) => {
-            *channel = channel.wrapping_add(STORAGE_ID_OFFSET);
-            // Note: source/target in Messages are *worker* IDs, not operator IDs.
-            // We leave them as-is.
-        }
-        TimelyEvent::PushProgress(e) => {
-            e.op_id = e.op_id.wrapping_add(STORAGE_ID_OFFSET);
-        }
-        TimelyEvent::CommChannels(e) => {
-            e.identifier = e.identifier.wrapping_add(STORAGE_ID_OFFSET);
-        }
-        TimelyEvent::Park(_) | TimelyEvent::Text(_) => {
-            // No IDs to remap.
-        }
     }
 }
